Feature/simple partial apply (#14)
* refactored PartialApplicationBuilder

* used just int

* refactored PartialApplicationBuilder

* minor changes

* defined <DebugType>full</DebugType>

* updated md files

* new build

* fixed typos

* removed redunduncy for partial extentions

* removed generated code

* changed extention of generated code

* minor

* fix build

* back generated code

* new build

* fixing build

* fixing build

* fixing build

* added .codecov.yml

+++ SCurry.Builders.Tests/CommonBuilderTests.cs
@@ -11,7 +11,7 @@ namespace SCurry.Builders.Tests
         [InlineData(0)]
         [InlineData(1)]
         [InlineData(3)]
-        public void TypeParameters_AppendResult_Test(ushort count)
+        public void TypeParameters_AppendResult_Test(int count)
         {
             var expected = count == 0
                 ? "TResult"
@@ -28,7 +28,7 @@ namespace SCurry.Builders.Tests
         [InlineData(0)]
         [InlineData(1)]
         [InlineData(3)]
-        public void TypeParameters_NoResult_Test(ushort count)
+        public void TypeParameters_NoResult_Test(int count)
         {
             var expected = count == 0
                 ? string.Empty
